requirements.txt
Flask==3.0.3
requests==2.31.0
gunicorn==22.0.0
cohere>=5.0.0
chromadb-client>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
